Change user authentification, add join and leave actions, fix pagination, restructure views
User authentication uses a cookie for now because the User.Identity.IsAuthenticated method returns false when reopening the browser window. Its set to true when the Google API is called, but that would also trigger the login form which is not wanted because then a new user would never see the start page.

Page indexes are now limited to the total number of pages. When you request a page index smaller than 1, page 1 is returned. When you request a page index bigger than the total number of pages, the max page is returned. This also fixes the problem that an empty table is shown when you delete all groups on the current page.